talentscout-chatbot/talentscout-chatbot/requirements.txt
streamlit>=1.36.0
openai>=1.35.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pandas>=2.2.2
